midi/midi.go
/*
The Midi package defines utilities for reading and writing midi sound files.
MIDI Files contain one or more MIDI streams, with time information for each
event. Song, sequence, and track structures, tempo and time signature
information, are all supported. Track names and other descriptive information
may be stored with the MIDI data.
*/
package midi

import (
	"bytes"
	"encoding/binary"
	"fmt"
	"io"
)

const (
	HeaderSizeError = "expected a header length of 16 but found a length of %v"
)

/*
The EventProcessor interface provides an API for parsing bytes out of a MIDI
file to construct a TrackEvent. At its core, each EventProcessor should be able
to create a fully constructed TrackEvent type. Because EventProcessors are
created by factories that have "claimed" the current event, if there is a
failure in parsing, the EventProcessor should return a non-nil error.
*/
type EventProcessor interface {
	Process(reader io.ByteReader) error
}

/*
The EventFactory interface provides a factory API for constructing MIDI
EventProcessors. EventFactory instances are registered with a MidiParser, and
are given a chance to process incoming events by returning an EventProcessor
when its ConstructProcessor() method is called. Returning a nil EventProcessor
indicates the event cannot be processed by this EventFactory's EventProcessor.

Whether or not an EventFactory can construct an EventProcessor for a given event
is based on the value of the first byte of a TrackEvent. An error will be raised
by the MidiParser if more than one EventFactory is registered as being able to
handle a specific byte.
*/
type EventFactory interface {
	ConstructProcessor(byte) EventProcessor
}

/*
isLastByte returns true when the passed in byte is the last in a variable length
quanitity.
*/
func isLastByte(b *byte) bool {
	return *b&msbMask != msbMask
}

/*
ReadVariableLengthQuantity consumes bytes from a io.Reader according to the
variable length quantity format, where each byte in the sequence, except the
last, has a 1 in the most significant bit. It returns a uint64 containing the
value of the sequence.
*/
func ReadVariableLengthQuantity(reader io.ByteReader) uint64 {
	value := uint64(0)
	current, err := reader.ReadByte()
	for i := 0; i < 8 && err == nil; i++ {
		value = value<<7 + uint64(current&sevenBitMask)
		if isLastByte(&current) {
			break
		}
		current, err = reader.ReadByte()
	}
	return value
}

/*
UnmarshalBinary reads in bytes from data and populates the Midi receiver. This
method satisfies the encoder.BinaryUnmarshaler interface.
*/
func (m *Midi) UnmarshalBinary(data []byte) error {
	buffer := bytes.NewBuffer(data)
	if err := m.unmarshalHeaderChunk(buffer); err != nil {
		return err
	}
	return nil
}

/*
The unmarshalHeaderChunk method parses out a Midi header chunk. If there is
an error parsing out a valid header chunk, a non-nil error is returned.
*/
func (m *Midi) unmarshalHeaderChunk(reader io.Reader) error {
	var chunk Chunk
	if err := binary.Read(reader, binary.BigEndian, &chunk); err != nil {
		return err
	}
	if chunk.Length != uint32(16) {
		return fmt.Errorf(HeaderSizeError, chunk.Length)
	}
	var format, ntrks, division uint16
	if err := binary.Read(reader, binary.BigEndian, &format); err != nil {
		return err
	}
	if err := binary.Read(reader, binary.BigEndian, &ntrks); err != nil {
		return err
	}
	if err := binary.Read(reader, binary.BigEndian, &division); err != nil {
		return err
	}
	m.HeaderChunk = &HeaderChunk{
		Chunk:    &chunk,
		Format:   format,
		Ntrks:    ntrks,
		Division: division,
	}
	return nil
}

/*
MidiEventFactory creates MidiEventProcessors when channel voice messages are
encountered. Given a byte, this factory will inspect the 4 high-order bits to
determine whether they match any of the known channel voice message events.
*/
type midiEventFactory struct {
	last EventProcessor
}

func (*midiEventFactory) ConstructProcessor(midiByte byte) EventProcessor {
	switch midiByte & highOrderMask {
	case NoteOffEvent:
	case NoteOnEvent:
	case PolyphonicKeyPressure:
	case ControlChange:
	case ProgramChange:
	case ChannelPressure:
	case PitchWheelChange:
	}
	return nil
}
